// apps/api/src/auth/dto.d.ts
export declare class SignupDto {
    email: string;
    password: string;
    name: string;
    orgName: string;
}
export declare class LoginDto {
    email: string;
    password: string;
}
export declare class SwitchOrganizationDto {
    organizationId: string;
}
export declare class AuthResponseDto {
    accessToken: string;
}